fix(incremental): Gate prefix-sensitive scanner reuse after edits
- Add a behavior-based gate for prefix-sensitive scanner reuse after edits.
- Keep certified checkpoint reuse for scanners without the new requirement.
- Track compact checkpoint transfer proof and expand regression coverage and documentation.

Buckley-Change-Hash: sha256:7e33b235e6e970f67c8dfc5bf4ee486d22febcd76202b24d5daa2aeaeb354d66
Buckley-Change-Stats: files=11 insertions=94 deletions=63 binaries=0

+++ incremental.go
@@ -83,7 +83,7 @@ type reuseCursor struct {
 	// candidates that have exact scanner bytes but no proof that the recorded
 	// reduction owned the live parser frontier. Leaves use their independent
 	// shift and checkpoint checks; non-leaves stay barred until that proof exists.
-	rejectFrontierProofUnavailable uint64
+	rejectFrontierProofUnavailable uint32
 	forestFastPath                 bool
 	// compactMaterialized marks old trees whose node states came from compact
 	// replay. Checkpoint snapshots authenticate scanner state, but they do not
